Ignore everything after '||' when normalizing
For our use, part of a true callnumber collation key is going to
be the record identifier. The collation keys will now ignore
anything after '||'.

* <callnumber>||bibID becomes <callnumber_normalized>!!bib_id

The '!!' is used because it sorts before all letters. It's _not_
used as the delimiter in the input because it's not impossible to
believe it'll show up in a callnumber field.